Rename IndexedBmpWriter.h to BitmapFile.h
