+++ OP2Utility.vcxproj
@@ -126,7 +126,7 @@
     <ClInclude Include="src\Bitmaps\BmpHeader.h" />
     <ClInclude Include="src\Bitmaps\Color.h" />
     <ClInclude Include="src\Bitmaps\ImageHeader.h" />
-    <ClInclude Include="src\Bitmaps\IndexedBmpWriter.h" />
+    <ClInclude Include="src\Bitmaps\BitmapFile.h" />
     <ClInclude Include="src\Maps\CellType.h" />
     <ClInclude Include="src\Maps\MapData.h" />
     <ClInclude Include="src\Maps\MapHeader.h" />
