+++ src/Bitmaps/BitmapFile.h
